fix: round-1 review — JSON-fallback coercion, dup-name/float-only/zero-field refusals, null guards (#44)
Six fix families from the PR #49 round-1 review:

1. Integral JSON-fallback scalars vs f32 view fields (all three): the
   JSON route types whole-number floats as int-class tokens (the host's
   serializer AND our own fallback encoder spell 1.0 as '1'), which the
   f32 view setters skipped — the graceful-degrade path broke on whole
   numbers. The decode layers now coerce across numeric classes exactly
   where the host's own JSON parse would: int classes always land in an
   f32 field (host-parser rounding); a FLOAT class lands in an int
   field only when EXACT (finite, integral, in range — mirroring the
   packed SET refusal rules), skipped otherwise. Rust: PackedField's
   from_scalar matrix; crystal: the macro set_field arms + coerce_f32_*
   helpers; csharp: new Scalar.TryF32/TryI64/TryU64/TryBool accessors
   the views assign through (docs updated). Tests: '{"a":2}' through
   the fallback + whole-number set_from round trips, per language.

2. Duplicate component names in the typed 2-component batch (all
   three): the stream would carry two copies per entity and the
   positional write-back would let the unchanged copy overwrite the
   other's writes — refused before any host call (the ruby block tier's
   duplicate-FIELD refusal, one level up). Rust:
   BatchError::DuplicateComponent (names are runtime strings — checked
   over all pairs in batch_core); crystal: ArgumentError; csharp:
   ArgumentException. Raw tiers stay script-owned. Tests per language.

3. C# bool batch fields: chose the FLOAT-ONLY restriction over
   per-field marshaling — the zero-copy in-place reinterpret is the
   design goal, and it is the same restriction rust enforces at compile
   time (no non-float macro arm). Enforced by a cached per-type
   reflection check (ViewInfo<T>): bool/int fields, padding and
   zero-field structs refuse with a pointed BatchRefusedException
   before any host call. Docs drop the bools-as-0/1 claim. Test.

4. Rust zero-stride views: batch_view! now requires at least one field
   (the matcher's ),+ — no zero-field arm, the same mechanism as the
   no-int-type arm), so 'Marker {}' is a compile error instead of a
   chunks_exact_mut(0) panic. Crystal gets the macro-raise twin; csharp
   is covered by ViewInfo's zero-field refusal.

5. C# null guards: ArgumentNullException.ThrowIfNull on
   GetInto(view, scratch) / SetFrom(view) / BatchGet / BatchSet and the
   Batch/BatchWhile delegates.

6. Example field alignment: Boid/BoidVel declare 'float' (was
   'double'). Verified against Declare.cs and the generated registry:
   double|float BOTH map to schema "f32" (the vocabulary has no f64 —
   'double' exists only so non-exact defaults format at full precision
   before %.14g), so the generated component was f32 all along; 'float'
   keeps the declaration visibly aligned with the float-typed views.

zig build test green (all seven binaries); all three example games
re-verified end-to-end (exact transcripts, generated Boid fields
confirmed f32).

Refs #44
Refs #41

+++ examples/csharp-game/components/Boid.cs
@@ -6,9 +6,14 @@
 // field-name choice: script-declared components codegen with fields SORTED
 // BY NAME, and the batch stream walks the generated struct's field order —
 // x < y keeps the declared order and the stream order identical.
+// `float` fields (not `double` like Hunger's): both spell schema "f32" —
+// the declare vocabulary has no f64, and `double` exists only so
+// non-exact DEFAULTS format at full precision before %.14g — but the
+// batch views these fields back are `float`-typed, so `float` here keeps
+// the declaration visibly aligned with the view (0.0 is exact either way).
 [LabelleComponent]
 record Boid
 {
-    public double x = 0.0;
-    public double y = 0.0;
+    public float x = 0.0f;
+    public float y = 0.0f;
 }
